Create Stub, Profile page, sign up, and login updates
Create stub only the page with no functionality. Profile page won't display user data. Sign up should be working. Login works, but there is a massive flaw in it. A user with one username can enter in a completely different password and still get in so long as the username and password are in the database. Maybe merge user and user password classes